Refactor empty view to simplify and add filter name in the empty view message.
